+++ src/ContainerImpl.test.ts
@@ -1,4 +1,4 @@
-import { container, ContainerImpl } from './ContainerImpl';
+import { container } from './ContainerImpl';
 import { dependencies, singleton, inject } from './decorators';
 
 describe('register type', () => {
@@ -290,99 +290,6 @@ describe('nested resolve', () => {
     const list = service.myList();
     expect(list).toEqual([
       'default',
-      'service entity',
-    ]);
-  });
-});
-
-describe('nested containers', () => {
-  abstract class Service {
-    abstract some(): string;
-  }
-
-  class Service1 extends Service {
-    some(): string {
-      return 'service 1';
-    }
-  }
-
-  class Service2 extends Service {
-    some(): string {
-      return 'service 2';
-    }
-  }
-
-  @dependencies(Service)
-  class Consumer {
-    constructor(private service: Service) {
-    }
-
-    do() {
-      return this.service.some();
-    }
-  }
-
-  const container1 = new ContainerImpl();
-  const container2 = new ContainerImpl(container1);
-  const container3 = new ContainerImpl(container2);
-
-  container1.registerType(Service1).as(Service);
-
-  test('container 1 -> 3', () => {
-    const consumer = container3.resolve(Consumer);
-    expect(consumer instanceof Consumer).toBe(true);
-    expect(consumer!.do()).toBe('service 1');
-  });
-
-  test('container 2 overrides container 1', () => {
-    container2.registerType(Service2).as(Service);
-    const consumer = container3.resolve(Consumer);
-    expect(consumer instanceof Consumer).toBe(true);
-    expect(consumer!.do()).toBe('service 2');
-  });
-
-  test('singleton resolving', () => {
-    @singleton
-    class Database {
-      readonly entities: string[];
-
-      constructor() {
-        this.entities = ['entity 1', 'entity 2'];
-      }
-    }
-
-    @dependencies(Database)
-    class Repository {
-      private db: Database;
-
-      constructor(db: Database) {
-        this.db = db;
-      }
-
-      list() {
-        return this.db.entities;
-      }
-    }
-
-    @dependencies(Repository)
-    class Service {
-      private repository: Repository;
-
-      constructor(repository: Repository) {
-        this.repository = repository;
-      }
-
-      myList() {
-        const entities = this.repository.list();
-        return entities.concat('service entity');
-      }
-    }
-
-    const service = container.resolve(Service)!;
-    const list = service.myList();
-    expect(list).toEqual([
-      'entity 1',
-      'entity 2',
       'service entity',
     ]);
   });